updated math/rand to crypto/rand

+++ sortingAlgs/quickSort/quickSort_test.go
@@ -7,13 +7,13 @@ import (
 )
 
 func TestInvertedSort(t *testing.T) {
-	arr := []int{5, 4, 3, 2, 1}
+	arr := []byte{5, 4, 3, 2, 1}
 	quickSort(arr, 0, len(arr)-1)
 	for i := 0; i < len(arr)-1; i++ {
 		if arr[i] > arr[i+1] {
 			for j := range arr {
 				val := arr[j]
-				fmt.Print(strconv.Itoa(val))
+				fmt.Print(strconv.Itoa(int(val)))
 			}
 			t.Errorf("%d > %d", arr[i], arr[i+1])
 		}
@@ -21,13 +21,13 @@ func TestInvertedSort(t *testing.T) {
 }
 
 func TestSortedSort(t *testing.T) {
-	arr := []int{1, 2, 3, 4, 5, 6, 7, 8, 9, 10}
+	arr := []byte{1, 2, 3, 4, 5, 6, 7, 8, 9, 10}
 	quickSort(arr, 0, len(arr)-1)
 	for i := 0; i < len(arr)-1; i++ {
 		if arr[i] > arr[i+1] {
 			for j := range arr {
 				val := arr[j]
-				fmt.Print(strconv.Itoa(val))
+				fmt.Print(strconv.Itoa(int(val)))
 			}
 			t.Errorf("%d > %d", arr[i], arr[i+1])
 		}
